+++ Cesium.externs.js
@@ -144,6 +144,11 @@ Cesium.Color.unpack = function(color) {};
 Cesium.Color.byteToFloat = function(component) {};
 
 /**
+ * @type {!Cesium.Color}
+ */
+Cesium.Color.WHITE;
+
+/**
  * @constructor
  * @param {string} text
  * @param {Object=} opt_description
@@ -245,6 +250,7 @@ Cesium.BillboardCollection = function(opt_options) {};
  *   verticalOrigin: (Cesium.VerticalOrigin|undefined),
  *   horizontalOrigin: (Cesium.HorizontalOrigin|undefined),
  *   pixelOffsetScaleByDistance : (Cesium.NearFarScalar|undefined),
+ *   scale: (number|undefined),
  *   scaleByDistance: (Cesium.NearFarScalar|undefined),
  *   position: !Cesium.Cartesian3,
  *   scale : (number|undefined)
@@ -422,7 +428,7 @@ Cesium.Camera.prototype.constrainedAxisAngle;
 
 
 /**
- * @type {Cesium.Cartographic} .
+ * @type {!Cesium.Cartographic} .
  */
 Cesium.Camera.prototype.positionCartographic;
 
@@ -456,6 +462,16 @@ Cesium.Camera.prototype.pitch;
  * @const
  */
 Cesium.Camera.prototype.roll;
+
+/**
+ * @type {Cesium.Event}
+ */
+Cesium.Camera.prototype.moveStart;
+
+/**
+ * @type {Cesium.Event}
+ */
+Cesium.Camera.prototype.moveEnd;
 
 
 /**
@@ -628,7 +644,7 @@ Cesium.CameraEventAggregator.prototype.isDestroyed = function() {};
 
 
 /**
-*/
+ */
 Cesium.CameraEventAggregator.prototype.destroy = function() {};
 
 
@@ -640,11 +656,22 @@ Cesium.Camera.prototype.setPositionCartographic;
 
 /**
  * @typedef {{
- *  position: (Cesium.Cartesian3|undefined),
- *  positionCartographic: (Cesium.Cartographic|undefined),
- *  heading: (number|undefined),
- *  pitch: (number|undefined),
- *  roll: (number|undefined)
+ *    heading: (number|undefined),
+ *    pitch: (number|undefined),
+ *    roll: (number|undefined)
+ * }}
+ */
+Cesium.optionsOrientation;
+
+/**
+ * @typedef {{
+ *  destination: (Cesium.Cartesian3|Cesium.Rectangle|undefined),
+ *  orientation: (Cesium.optionsOrientation|undefined),
+ *  position: (Cesium.RemovedAPI|undefined),
+ *  positionCartographic: (Cesium.RemovedAPI|undefined),
+ *  heading: (undefined|Cesium.RemovedAPI),
+ *  pitch: (undefined|Cesium.RemovedAPI),
+ *  roll: (undefined|Cesium.RemovedAPI)
  * }}
  */
 Cesium.optionsCameraSetView;
@@ -672,6 +699,31 @@ Cesium.Camera.prototype.lookAt = function(eye, target, up) {};
  */
 Cesium.Camera.prototype.lookAtTransform = function(transform) {};
 
+/**
+ * @param {Cesium.Cartesian3} axis
+ * @param {number} angle
+ */
+Cesium.Camera.prototype.rotate = function(axis, angle) {};
+
+/**
+ * @param {number} angle
+ */
+Cesium.Camera.prototype.rotateUp = function(angle) {};
+
+/**
+ * @param {number} angle
+ */
+Cesium.Camera.prototype.rotateDown = function(angle) {};
+
+/**
+ * @param {number} angle
+ */
+Cesium.Camera.prototype.rotateLeft = function(angle) {};
+
+/**
+ * @param {number} angle
+ */
+Cesium.Camera.prototype.rotateRight = function(angle) {};
 
 /**
  * @param {number} amount .
@@ -745,6 +797,25 @@ Cesium.Camera.prototype.moveUp = function(amount) {};
  */
 Cesium.Camera.prototype.pickEllipsoid = function(windowPos, opt_ellipsoid) {};
 
+/**
+ * @typedef {{
+ *   destination: (!Cesium.Cartesian3|Cesium.Rectangle),
+ *   orientation: (!Cesium.optionsOrientation|undefined),
+ *   duration: (number|undefined),
+ *   complete: (function()|undefined),
+ *   cancel: (function()|undefined),
+ *   endTransform: (Cesium.Matrix4|undefined),
+ *   convert: (boolean|undefined),
+ *   maximumHeight: (number|undefined),
+ *   easingFunction: (function(number): number|undefined)
+ * }}
+ */
+Cesium.optionsCameraFlyTo;
+
+/**
+ * @param {!Cesium.optionsCameraFlyTo} options
+ */
+Cesium.Camera.prototype.flyTo = function(options) {};
 
 /**
  * @constructor
@@ -1023,6 +1094,12 @@ Cesium.Globe = function(ellipsoid) {};
  * @type {Cesium.Ellipsoid}
  */
 Cesium.Globe.prototype.ellipsoid;
+
+
+/**
+ * @type {!Cesium.Color}
+ */
+Cesium.Globe.prototype.baseColor;
 
 
 /**
@@ -1673,6 +1750,12 @@ Cesium.DefaultProxy = function(proxy) {};
  */
 Cesium.Event = function() {};
 
+/**
+ * @param {function(...)} listener
+ * @param {Object=} opt_scope
+ * @return {function()}
+ */
+Cesium.Event.prototype.addEventListener = function(listener, opt_scope) {};
 
 /**
  * @param {Function} listener
@@ -2156,33 +2239,6 @@ Cesium.Rectangle.fromCartographicArray = function(cartographics, opt_result) {};
 
 
 /**
- * @typedef {{asynchronous: (boolean|undefined),
- *     height: (number|undefined),
- *     rectangle: (Cesium.Rectangle|undefined),
- *     material: (Cesium.Material|undefined)}}
- */
-Cesium.RectanglePrimitiveOptions;
-
-/**
- * @constructor
- * @param {Cesium.RectanglePrimitiveOptions} options .
- * @extends {Cesium.Primitive}
- */
-Cesium.RectanglePrimitive = function(options) {};
-
-
-/**
- * @type {!Cesium.Material} material .
- */
-Cesium.RectanglePrimitive.prototype.material;
-
-/**
- * @type {!Cesium.Rectangle} rectangle .
- */
-Cesium.RectanglePrimitive.prototype.rectangle;
-
-
-/**
  * @constructor
  */
 Cesium.FeatureDetection = function() {};
@@ -2297,6 +2353,14 @@ Cesium.Matrix4 = function(opt_a00, opt_a10, opt_a20, opt_a30,
 
 
 /**
+ * @param {Cesium.Matrix4} matrix
+ * @param {Cesium.Matrix4=} opt_result
+ * @return {!Cesium.Matrix4}
+ */
+Cesium.Matrix4.clone = function(matrix, opt_result) {};
+
+
+/**
  * @param {Cesium.Cartesian3} translation .
  * @param {Cesium.Matrix4=} opt_result .
  * @return {!Cesium.Matrix4} .
@@ -2314,10 +2378,27 @@ Cesium.Matrix4.multiply = function(left, right, result) {};
 
 
 /**
+ * @param {Cesium.Matrix4|undefined} matrix1
+ * @param {Cesium.Matrix4|undefined} matrix2
+ * @param {number} epsilon
+ * @return {boolean}
+ */
+Cesium.Matrix4.equalsEpsilon = function(matrix1, matrix2, epsilon) {};
+
+
+/**
  * @param {Cesium.Matrix4} matrix .
  * @return {boolean} .
  */
 Cesium.Matrix4.prototype.equals = function(matrix) {};
+
+
+/**
+ * @param {Cesium.Matrix4} matrix
+ * @param {number} epsilon
+ * @return {boolean}
+ */
+Cesium.Matrix4.prototype.equalsEpsilon = function(matrix, epsilon) {};
 
 
 /**
@@ -2349,35 +2430,6 @@ Cesium.Matrix4.multiplyByVector = function(matrix, point, result) {};
  * @return {Array.<number>} .
  */
 Cesium.Matrix4.toArray = function(matrix) {};
-
-
-/**
- * @constructor
- * @param {{url: (string|undefined),
- *          fileExtension: (string|undefined),
- *          proxy: (Object|undefined),
- *          rectangle: (Cesium.Rectangle|undefined),
- *          minimumLevel: (number|undefined),
- *          maximumLevel: (number|undefined),
- *          credit: (Cesium.Credit|string|undefined)}=} opt_opts
- * @extends {Cesium.ImageryProvider}
- */
-Cesium.OpenStreetMapImageryProvider = function(opt_opts) {};
-
-
-
-/**
- * @constructor
- * @param {{url: string,
- *          layers: string,
- *          parameters: (Object|undefined),
- *          rectangle: (Cesium.Rectangle|undefined),
- *          maximumLevel: (number|undefined),
- *          credit: (Cesium.Credit|string|undefined),
- *          proxy: (Object|undefined)}} options
- * @extends {Cesium.ImageryProvider}
- */
-Cesium.WebMapServiceImageryProvider = function(options) {};
 
 
 
@@ -2424,12 +2476,13 @@ Cesium.PerspectiveFrustrum.prototype.projectionMatrix;
 
 
 /**
- * @param {!Cesium.Cartesian2} canvasDim .
- * @param {number=} opt_dist .
+ * @param {!number} drawingBufferWidth
+ * @param {!number} drawingBufferHeight
+ * @param {!number} dist
+ * @param {!Cesium.Cartesian2} result
  * @return {!Cesium.Cartesian2}
  */
-Cesium.PerspectiveFrustrum.prototype.getPixelSize = function(canvasDim, opt_dist) {};
-
+Cesium.PerspectiveFrustrum.prototype.getPixelDimensions = function(drawingBufferWidth, drawingBufferHeight, dist, result) {};
 
 
 /**
@@ -2458,6 +2511,25 @@ Cesium.Context = function() {};
  */
 Cesium.Context.prototype.uniformState;
 
+/**
+ * @constructor
+ */
+Cesium.Fog = function() {};
+
+/**
+ * @type {boolean}
+ */
+Cesium.Fog.prototype.enabled;
+
+/**
+ * @type {number}
+ */
+Cesium.Fog.prototype.density;
+
+/**
+ * @type {number}
+ */
+Cesium.Fog.prototype.screenSpaceErrorFactor;
 
 
 /**
@@ -2478,9 +2550,21 @@ Cesium.Scene.prototype.canvas;
 
 
 /**
+ * @type {!Cesium.Color}
+ */
+Cesium.Scene.prototype.backgroundColor;
+
+
+/**
  * @type {!Cesium.Context}
  */
 Cesium.Scene.prototype.context;
+
+
+/**
+ * @type {!Cesium.Fog}
+ */
+Cesium.Scene.prototype.fog;
 
 
 /**
@@ -2489,8 +2573,19 @@ Cesium.Scene.prototype.initializeFrame = function() {};
 
 
 /**
+ * @param {Cesium.JulianDate=} opt_date
  */
-Cesium.Scene.prototype.render = function() {};
+Cesium.Scene.prototype.render = function(opt_date) {};
+
+/**
+ * @type {Cesium.Event}
+ */
+Cesium.Scene.prototype.preRender;
+
+/**
+ * @type {Cesium.Event}
+ */
+Cesium.Scene.prototype.postRender;
 
 
 /**
@@ -2561,6 +2656,98 @@ Cesium.Scene.prototype.skyAtmosphere;
  * @type {number}
  */
 Cesium.Scene.prototype.maximumAliasedLineWidth;
+
+
+/**
+ * @constructor
+ */
+Cesium.JulianDate = function() {};
+
+
+/**
+ * @return {Cesium.JulianDate}
+ */
+Cesium.JulianDate.now = function() {};
+
+
+/**
+ * @constructor
+ */
+Cesium.DataSource = function() {};
+
+
+/**
+ * @constructor
+ */
+Cesium.DataSourceCollection = function() {};
+
+/**
+ * @param {Cesium.DataSource|Promise.<Cesium.DataSource>} dataSource A data source or a promise to a data source to add to the collection.
+ *                                        When passing a promise, the data source will not actually be added
+ *                                        to the collection until the promise resolves successfully.
+ * @returns {Promise.<Cesium.DataSource>} A Promise that resolves once the data source has been added to the collection.
+ */
+Cesium.DataSourceCollection.prototype.add = function(dataSource) {};
+
+/**
+ * Removes a data source from this collection, if present.
+ *
+ * @param {Cesium.DataSource} dataSource The data source to remove.
+ * @param {Boolean} [destroy=false] Whether to destroy the data source in addition to removing it.
+ * @returns {Boolean} true if the data source was in the collection and was removed,
+ *                    false if the data source was not in the collection.
+ */
+Cesium.DataSourceCollection.prototype.remove = function(dataSource, destroy) {};
+
+
+/**
+ * @constructor
+ * @param {{scene: Cesium.Scene,
+            dataSourceCollection: Cesium.DataSourceCollection}} opt_opts
+ */
+Cesium.DataSourceDisplay = function(opt_opts) {};
+
+
+/**
+ * @param {Cesium.JulianDate} time The simulation time.
+ * @returns {Boolean} True if all data sources are ready to be displayed, false otherwise.
+ */
+Cesium.DataSourceDisplay.prototype.update = function(time) {};
+
+
+/**
+ * @type {Cesium.CustomDataSource}
+ */
+Cesium.DataSourceDisplay.prototype.defaultDataSource;
+
+
+
+/**
+ * @param {string} name
+ * @constructor
+ */
+Cesium.CustomDataSource = function(name) {};
+
+
+/**
+ * @type {Cesium.EntityCollection}
+ */
+Cesium.CustomDataSource.prototype.entities;
+
+
+
+/**
+ * @constructor
+ */
+Cesium.EntityCollection = function() {}
+
+
+/**
+ * @type {!Cesium.UniformState}
+ */
+Cesium.Context.prototype.uniformState;
+
+
 
 /**
  * @type {boolean}
@@ -2711,7 +2898,7 @@ Cesium.ScreenSpaceCameraController.prototype.inertiaSpin;
 /**
  * @type {number}
  */
-Cesium.ScreenSpaceCameraController.prototype.ineartiaTranslate;
+Cesium.ScreenSpaceCameraController.prototype.inertiaTranslate;
 
 /**
  * @type {number}
@@ -3065,53 +3252,104 @@ Cesium.HeightReference.RELATIVE_TO_GROUND;
 Cesium.WebMapTileServiceImageryProvider = function(options) {};
 
 
-
 /**
  * @typedef {{
  *   url: string,
- *   layer: string,
  *   format: string,
+ *   layer: string,
  *   style: string,
  *   tileMatrixSetID: string,
+ *   tileWidth: (number|undefined),
+ *   tileHeight: (number|undefined),
  *   tilingScheme: (Cesium.TilingScheme|undefined),
- *   maximumLevel: (number|undefined)
+ *   proxy: (Object|undefined),
+ *   rectangle: (Cesium.Rectangle|undefined),
+ *   minimumLevel: (number|undefined),
+ *   maximumLevel: (number|undefined),
+ *   credit: (Cesium.Credit|string|undefined),
+ *   subdomains: (string|Array.<string>|undefined)
  * }}
  */
 Cesium.WebMapTileServiceImageryProviderOptions;
 
 
+
 /**
- * @type {number}
+ * @param {!Cesium.UrlTemplateImageryProviderOptions} options
+ * @extends {Cesium.ImageryProvider}
+ * @constructor
  */
-Cesium.WebMapTileServiceImageryProviderOptions.prototype.maximumLevel;
+Cesium.UrlTemplateImageryProvider = function(options) {};
 
 
 /**
- * @type {Cesium.TilingScheme}
+ * @typedef {{
+ *   url: string,
+ *   subdomains: (string|Array.<string>|undefined),
+ *   proxy: (Object|undefined),
+ *   credit: (Cesium.Credit|string|undefined),
+ *   minimumLevel: (number|undefined),
+ *   maximumLevel: (number|undefined),
+ *   rectangle: (Cesium.Rectangle|undefined),
+ *   tilingScheme: (Cesium.TilingScheme|undefined),
+ *   tileWidth: (number|undefined),
+ *   tileHeight: (number|undefined),
+ *   hasAlphaChannel: (boolean|undefined)
+ * }}
  */
-Cesium.WebMapTileServiceImageryProviderOptions.prototype.tilingScheme;
+Cesium.UrlTemplateImageryProviderOptions;
 
 
 /**
- * @type {string}
+ * @typedef {{
+ *   url: string,
+ *   layers: string,
+ *   parameters: (Object|undefined),
+ *   rectangle: (Cesium.Rectangle|undefined),
+ *   tilingScheme: (Cesium.TilingScheme|undefined),
+ *   tileWidth: (number|undefined),
+ *   tileHeight: (number|undefined),
+ *   minimumLevel: (number|undefined),
+ *   maximumLevel: (number|undefined),
+ *   credit: (Cesium.Credit|string|undefined),
+ *   proxy: (Object|undefined),
+ *   subdomains: (string|Array.<string>|undefined)
+ * }}
  */
-Cesium.WebMapTileServiceImageryProviderOptions.prototype.tileMatrixSetID;
+Cesium.WebMapServiceImageryProviderOptions;
+
 
 
 /**
- * @type {string}
+ * @constructor
+ * @param {Cesium.WebMapServiceImageryProviderOptions} options
+ * @extends {Cesium.ImageryProvider}
  */
-Cesium.WebMapTileServiceImageryProviderOptions.prototype.format;
+Cesium.WebMapServiceImageryProvider = function(options) {};
 
 
 /**
- * @type {string}
+ * @type {function(Object=)}
  */
-Cesium.WebMapTileServiceImageryProviderOptions.prototype.layer;
+Cesium.loadWithXhr;
 
 
 /**
- * @type {string}
+ * @type {function(...)}
+ */
+Cesium.loadWithXhr.load;
+
+
+/**
+ * @param {string} workerName
+ * @param {number=} opt_maximumActiveTasks
+ * @constructor
+ */
+Cesium.TaskProcessor = function(workerName, opt_maximumActiveTasks) {};
+
+
+/**
+ * @return {boolean}
  */
 Cesium.WebMapTileServiceImageryProviderOptions.prototype.url;
 
@@ -3120,3 +3358,5 @@ Cesium.WebMapTileServiceImageryProviderOptions.prototype.url;
  * @constructor
  */
 Cesium.Cesium3DTileset = function(options){};
+
+Cesium.TaskProcessor.prototype.isDestroyed = function() {};