+++ examples/vectors.js
@@ -152,13 +152,14 @@ var vectorLayer = new ol.layer.Vector({
   style: styleFunction
 });
 
-
 var vectorSource2 = new ol.source.Vector({
   features: [iconFeature, textFeature, cervinFeature]
 });
-
-var vectorLayer2 = new ol.layer.Vector({
+var imageVectorSource = new ol.source.ImageVector({
   source: vectorSource2
+});
+var vectorLayer2 = new ol.layer.Image({
+  source: imageVectorSource
 });
 
 var dragAndDropInteraction = new ol.interaction.DragAndDrop({
@@ -171,15 +172,11 @@ var dragAndDropInteraction = new ol.interaction.DragAndDrop({
   ]
 });
 
-
 var map = new ol.Map({
   interactions: ol.interaction.defaults().extend([dragAndDropInteraction]),
   layers: [
     new ol.layer.Tile({
-      source: new ol.source.BingMaps({
-        key: 'Ak-dzM4wZjSqTlzveKz5u0d4IQ4bRzVI309GxmkgSVr1ewS6iPSrOvOKhA-CJlm3',
-        imagerySet: 'Aerial'
-      })
+      source: new ol.source.OSM()
     }),
     vectorLayer,
     vectorLayer2
@@ -214,9 +211,11 @@ dragAndDropInteraction.on('addfeatures', function(event) {
 var ol3d = new olcs.OLCesium({map: map, target: 'map3d'});
 var scene = ol3d.getCesiumScene();
 var terrainProvider = new Cesium.CesiumTerrainProvider({
-  url: '//assets.agi.com/stk-terrain/world'
+  url: '//assets.agi.com/stk-terrain/world',
+  requestVertexNormals: true
 });
 scene.terrainProvider = terrainProvider;
+scene.globe.enableLighting = true;
 ol3d.setEnabled(true);
 
 var csLabels = new Cesium.LabelCollection();
@@ -275,3 +274,5 @@ function toggleClampToGround() {
   map.addLayer(vectorLayer);
   map.addLayer(vectorLayer2);
 }
+
+ol3d.enableAutoRenderLoop();