+++ src/vectorsynchronizer.js
@@ -15,6 +15,7 @@ goog.require('olcs.core.VectorLayerCounterpart');
  * @constructor
  * @extends {olcs.AbstractSynchronizer.<olcs.core.VectorLayerCounterpart>}
  * @api
+ * @struct
  */
 olcs.VectorSynchronizer = function(map, scene, opt_converter) {
 
@@ -85,26 +86,39 @@ olcs.VectorSynchronizer.prototype.removeAllCesiumObjects = function(destroy) {
  */
 olcs.VectorSynchronizer.prototype.createSingleLayerCounterparts =
     function(olLayer) {
-  if (!(olLayer instanceof ol.layer.Vector)) {
+  if (!(olLayer instanceof ol.layer.Vector) &&
+      !(olLayer instanceof ol.layer.Image &&
+      olLayer.getSource() instanceof ol.source.ImageVector)) {
     return null;
   }
-  goog.asserts.assertInstanceof(olLayer, ol.layer.Vector);
+  goog.asserts.assertInstanceof(olLayer, ol.layer.Layer);
+
+  var source = olLayer.getSource();
+  if (source instanceof ol.source.ImageVector) {
+    source = source.getSource();
+  }
+
+  goog.asserts.assertInstanceof(source, ol.source.Vector);
   goog.asserts.assert(!goog.isNull(this.view));
 
   var view = this.view;
-  var source = olLayer.getSource();
   var featurePrimitiveMap = {};
   var counterpart = this.converter.olVectorLayerToCesium(olLayer, view,
       featurePrimitiveMap);
   var csPrimitives = counterpart.getRootPrimitive();
   var olListenKeys = counterpart.olListenKeys;
 
+  csPrimitives.show = olLayer.getVisible();
+
   olListenKeys.push(olLayer.on('change:visible', function(e) {
     csPrimitives.show = olLayer.getVisible();
   }));
 
   var onAddFeature = (function(feature) {
-    goog.asserts.assertInstanceof(olLayer, ol.layer.Vector);
+    goog.asserts.assert(
+        (olLayer instanceof ol.layer.Vector) ||
+        (olLayer instanceof ol.layer.Image)
+    );
     var context = counterpart.context;
     var prim = this.converter.convert(olLayer, view, feature, context);
     if (prim) {